# apps/runner/pkg/api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Daytona Runner API",
    "title": "Daytona Runner API",
    "contact": {},
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "v0.0.0-dev"
  },
  "paths": {
...
    "/sandboxes": {
      "post": {
        "description": "Create a sandbox",
        "produces": ["application/json"],
        "tags": ["sandbox"],
        "summary": "Create a sandbox",
        "operationId": "Create",
        "parameters": [
          {
            "description": "Create sandbox",
            "name": "sandbox",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateSandboxDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/StartSandboxResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sandboxes/{sandboxId}": {
      "get": {
        "description": "Get sandbox info",
        "produces": ["application/json"],
        "tags": ["sandbox"],
        "summary": "Get sandbox info",
        "operationId": "Info",
        "parameters": [
          {
            "type": "string",
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Sandbox info",
            "schema": {
              "$ref": "#/definitions/SandboxInfoResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sandboxes/{sandboxId}/backup": {
      "post": {
        "description": "Create sandbox backup",
        "produces": ["application/json"],
        "tags": ["sandbox"],
        "summary": "Create sandbox backup",
        "operationId": "CreateBackup",
        "parameters": [
          {
            "type": "string",
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path",
            "required": true
          },
          {
            "description": "Create backup",
            "name": "sandbox",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateBackupDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Backup started",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sandboxes/{sandboxId}/destroy": {
      "post": {
        "description": "Destroy sandbox",
        "produces": ["application/json"],
        "tags": ["sandbox"],
        "summary": "Destroy sandbox",
        "operationId": "Destroy",
        "parameters": [
          {
            "type": "string",
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Sandbox destroyed",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sandboxes/{sandboxId}/is-recoverable": {
      "post": {
        "description": "Check if the sandbox's error reason indicates a recoverable error",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["sandbox"],
        "summary": "Check if sandbox error is recoverable",
        "operationId": "IsRecoverable",
        "parameters": [
          {
            "type": "string",
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path",
            "required": true
          },
          {
            "description": "Error reason to check",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IsRecoverableDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/IsRecoverableResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sandboxes/{sandboxId}/network-settings": {
      "get": {
        "description": "Get sandbox network settings",
        "produces": ["application/json"],
        "tags": ["sandbox"],
        "summary": "Get sandbox network settings",
        "operationId": "GetNetworkSettings",
        "parameters": [
          {
            "type": "string",
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Network settings",
            "schema": {
              "$ref": "#/definitions/UpdateNetworkSettingsDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "description": "Update sandbox network settings",
        "produces": ["application/json"],
        "tags": ["sandbox"],
        "summary": "Update sandbox network settings",
        "operationId": "UpdateNetworkSettings",
        "parameters": [
          {
            "type": "string",
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path",
            "required": true
          },
          {
            "description": "Update network settings",
            "name": "sandbox",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateNetworkSettingsDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Network settings updated",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sandboxes/{sandboxId}/recover": {
      "post": {
        "description": "Recover sandbox from error state using specified recovery type",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["sandbox"],
        "summary": "Recover sandbox from error state",
        "operationId": "Recover",
        "parameters": [
          {
            "type": "string",
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path",
            "required": true
          },
          {
            "description": "Recovery parameters",
            "name": "recovery",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RecoverSandboxDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sandbox recovered",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/sandboxes/{sandboxId}/resize": {
      "post": {
        "description": "Resize sandbox",
        "produces": ["application/json"],
        "tags": ["sandbox"],
        "summary": "Resize sandbox",
        "operationId": "Resize",
        "parameters": [
          {
            "type": "string",
            "description": "Sandbox ID",
            "name": "sandboxId",
            "in": "path",
            "required": true
          },
          {
            "description": "Resize sandbox",
            "name": "sandbox",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ResizeSandboxDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sandbox resized",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateSandboxDTO": {
      "type": "object",
      "required": ["id", "osUser", "snapshot", "userId"],
      "properties": {
        "authToken": {
          "type": "string"
        },
        "cpuQuota": {
          "type": "integer",
          "minimum": 1
        },
        "domainAllowList": {
          "type": "string"
        },
        "entrypoint": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "env": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "fromVolumeId": {
          "type": "string"
        },
        "gpuQuota": {
          "type": "integer",
          "minimum": 0
        },
        "id": {
          "type": "string"
        },
        "linkedSandboxId": {
          "description": "LinkedSandboxId identifies an existing sandbox this sandbox should be co-located with.\nWhen set, the runner should attach both sandboxes to a shared local network so they can communicate.",
          "type": "string"
        },
        "memoryQuota": {
          "type": "integer",
          "minimum": 1
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "name": {
          "description": "Name is the human-readable sandbox name. When present, it is used as the\nnetwork alias for link networks so linked sandboxes can resolve each other\nby name.",
          "type": "string"
        },
        "networkAllowList": {
          "type": "string"
        },
        "networkBlockAll": {
          "type": "boolean"
        },
        "organizationId": {
          "description": "Nullable for backward compatibility",
          "type": "string"
        },
        "osUser": {
          "type": "string"
        },
        "otelEndpoint": {
          "type": "string"
        },
        "regionId": {
          "type": "string"
        },
        "registry": {
          "$ref": "#/definitions/RegistryDTO"
        },
        "sandboxClass": {
          "description": "Optional for backward compatibility, but when provided, indicates the class of sandbox to create.",
          "type": "string"
        },
        "secretsToken": {
          "description": "SecretsToken is a runner-only token used solely to resolve plaintext\nsecrets from the API (separate from AuthToken, which lives in the sandbox).",
          "type": "string"
        },
        "skipStart": {
          "type": "boolean"
        },
        "snapshot": {
          "type": "string"
        },
        "storageQuota": {
          "type": "integer",
          "minimum": 1
        },
        "userId": {
          "type": "string"
        },
        "volumes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/dto.VolumeDTO"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "Type \"Bearer\" followed by a space and an API token.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}